+++ cpu/native/mtd/mtd_native.c
@@ -53,7 +53,7 @@ static int _init(mtd_dev_t *dev)
     return 0;
 }
 
-static int _read(mtd_dev_t *dev, void *buff, uint32_t addr, uint32_t size)
+static int _read(mtd_dev_t *dev, void *buff, uint32_t addr, size_t size)
 {
     mtd_native_dev_t *_dev = (mtd_native_dev_t*) dev;
     size_t mtd_size = dev->sector_count * dev->pages_per_sector * dev->page_size;
@@ -75,7 +75,7 @@ static int _read(mtd_dev_t *dev, void *buff, uint32_t addr, uint32_t size)
     return size;
 }
 
-static int _write(mtd_dev_t *dev, const void *buff, uint32_t addr, uint32_t size)
+static int _write(mtd_dev_t *dev, const void *buff, uint32_t addr, size_t size)
 {
     mtd_native_dev_t *_dev = (mtd_native_dev_t*) dev;
     size_t mtd_size = dev->sector_count * dev->pages_per_sector * dev->page_size;
@@ -104,7 +104,7 @@ static int _write(mtd_dev_t *dev, const void *buff, uint32_t addr, uint32_t size
     return size;
 }
 
-static int _erase(mtd_dev_t *dev, uint32_t addr, uint32_t size)
+static int _erase(mtd_dev_t *dev, uint32_t addr, size_t size)
 {
     mtd_native_dev_t *_dev = (mtd_native_dev_t*) dev;
     size_t mtd_size = dev->sector_count * dev->pages_per_sector * dev->page_size;
