+++ tests/pkg_littlefs2/main.c
@@ -47,7 +47,7 @@ static int _init(mtd_dev_t *dev)
     return 0;
 }
 
-static int _read(mtd_dev_t *dev, void *buff, uint32_t addr, uint32_t size)
+static int _read(mtd_dev_t *dev, void *buff, uint32_t addr, size_t size)
 {
     (void)dev;
 
@@ -59,7 +59,7 @@ static int _read(mtd_dev_t *dev, void *buff, uint32_t addr, uint32_t size)
     return size;
 }
 
-static int _write(mtd_dev_t *dev, const void *buff, uint32_t addr, uint32_t size)
+static int _write(mtd_dev_t *dev, const void *buff, uint32_t addr, size_t size)
 {
     (void)dev;
 
@@ -74,7 +74,7 @@ static int _write(mtd_dev_t *dev, const void *buff, uint32_t addr, uint32_t size
     return size;
 }
 
-static int _erase(mtd_dev_t *dev, uint32_t addr, uint32_t size)
+static int _erase(mtd_dev_t *dev, uint32_t addr, size_t size)
 {
     (void)dev;
